Make inline images great again (if great = bigger)
Inline images are now capped at 100% width. To prevent images with
taller-than-they-are-wide aspect ratios from taking over, the image
height has been limited to 75% of the viewport height. This will retain
some context around the message, too.

Drawback: the width and height had to expressly be defined as "auto",
otherwise images could be dragged to resize them. While that's a
desirable trait, the aspect ratio would break when trying to drag an
image past its height limit. This disables image resizing by dragging!
Since images will be nice and large by default and clicking on an image
will open it in your browser, I'm cool with this behavior.